ace.5.hack (2)/ace.5.hack/ace.5.hack/backend/llm/googlechatbot.py
from langchain_google_genai import ChatGoogleGenerativeAI
from langchain_core.messages import SystemMessage, HumanMessage, AIMessage
from dotenv import load_dotenv
import os

load_dotenv()

# Gemini model
model = ChatGoogleGenerativeAI(
    model="gemini-1.5-flash",
    temperature=0.7
)

chat_history = [
    SystemMessage(
        content="You are a professional ai assistent who can hendle many task and give the suggestion"
    )
]

while True:
    user_input = input("You: ")

    if not user_input:
        print("AI: Please type something...")
        continue

    if user_input.lower() == "exit":
        break

    chat_history.append(HumanMessage(content=user_input))

    result = model.invoke(chat_history)

    chat_history.append(AIMessage(content=result.content))

    print("AI:", result.content)

print(chat_history)