ace.5.hack (2)/ace.5.hack/ace.5.hack/requirements.txt
fastapi
uvicorn[standard]
pydantic<3.0.0
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface-hub
sentence-transformers
faiss-cpu
python-dotenv
httpx
Pillow
chromadb
sse-starlette
aiofiles
